requirements.txt
ultralytics
opencv-python
deep-sort-realtime
numpy
pandas
pytorch
yt-dlp